src/requirement.txt
cryptography==41.0.5
dnspython==2.4.2
exchangelib==5.1.0
lxml==4.9.3
numpy==1.26.1
openpyxl==3.1.2
pandas==2.0.3
python-decouple==3.8
requests==2.31.0
requests-oauthlib==1.3.1
